# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Exclusive Job Portal API Documentation",
        "version": "0.1"
    },
    "paths": {
...
        "/api/subjects/": {
            "get": {
                "tags": [
                    "Subject"
                ],
                "summary": "Get subject details",
                "description": "Returns subject information",
                "operationId": "3bad08ee7681819d8ad4d5652c7d0793",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/subjects/questions/{subjectId}": {
            "get": {
                "tags": [
                    "Subject"
                ],
                "summary": "Get subjects details",
                "description": "Returns subjects's questions information",
                "operationId": "1fb663db87a332f8da1d768be5b7dc5c",
                "parameters": [
                    {
                        "name": "subjectId",
                        "in": "path",
                        "description": "Subject unique ID",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/subjects/chapters/{subjectId}": {
            "get": {
                "tags": [
                    "Subject"
                ],
                "summary": "Get chapters details",
                "description": "Returns chapter's questions information",
                "operationId": "b01edbc6c9be18840345512640d6eeae",
                "parameters": [
                    {
                        "name": "subjectId",
                        "in": "path",
                        "description": "Subject unique ID",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/subjects/chapters/questions/{chapterId}": {
            "get": {
                "tags": [
                    "Subject"
                ],
                "summary": "Get chapters questions details",
                "description": "Returns chapter's questions information",
                "operationId": "24ce346035af9ccef3e3487a77ed60a1",
                "parameters": [
                    {
                        "name": "chapterId",
                        "in": "path",
                        "description": "Chapter unique ID",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/subjects/questions-answers/score": {
            "post": {
                "tags": [
                    "Subject"
                ],
                "summary": "Store subject score",
                "description": "Returns score number",
                "operationId": "da1c3d589fba8943d8331efd8b20d4f7",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "parentId",
                                    "questions[]",
                                    "answers[]"
                                ],
                                "properties": {
                                    "parentId": {
                                        "type": "integer"
                                    },
                                    "questions[]": {
                                        "description": "Questions ID",
                                        "type": "array",
                                        "items": {
                                            "type": "int",
                                            "format": "id"
                                        },
                                        "collectionFormat": "multi"
                                    },
                                    "answers[]": {
                                        "description": "Answers ID",
                                        "type": "array",
                                        "items": {
                                            "type": "int",
                                            "format": "id"
                                        },
                                        "collectionFormat": "multi"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login Successful"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/faqs/": {
            "get": {
                "tags": [
                    "Utility"
                ],
                "summary": "Get faqs details",
                "description": "Returns faq information",
                "operationId": "effe94b6c7c41adc3fbc1d4d0a646162",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/abouts/": {
            "get": {
                "tags": [
                    "Utility"
                ],
                "summary": "Get abouts details",
                "description": "Returns about information",
                "operationId": "e6259e74a21c194a1c39da71424f1e61",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/videos/": {
            "get": {
                "tags": [
                    "Utility"
                ],
                "summary": "Get videos details",
                "description": "Returns video information",
                "operationId": "d0b9064c397567082993180251e79562",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    }
}